requirements.txt
customtkinter
opencv-python
Pillow